ENH: handle snapshots having >120 data-points
having >120 data-points means the snapshot included a
data-point intended for a later snapshot.
(and a later snapshot will be missing this data-point)

to account for this, we firt put data into an
intermediate map of arrs and only process data from it that has a timestamp
that falls into the current bucket time-window.

this makes it so data-points that end up in an early snapshot
get left in the intermediate map of arrs and put later into the correct
buckets.